+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 Features:
 
 Bug Fixes:
+* [RUBY-219](https://datastax-oss.atlassian.net/browse/RUBY-219) Sometimes get stack trace in metadata.rb due to failure in SortedSet initialization.
 
 
 # 3.0.0 GA